Before deleting an airplane, make sure no flight with the airplane is full

+++ src/main/java/com/utopia/flightservice/service/AirplaneService.java
@@ -1,10 +1,9 @@
 package com.utopia.flightservice.service;
 
-import java.util.List;
-
 import com.utopia.flightservice.entity.Airplane;
 import com.utopia.flightservice.exception.AirplaneNotFoundException;
 import com.utopia.flightservice.repository.AirplaneRepository;
+import com.utopia.flightservice.repository.FlightDao;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -13,25 +12,20 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
+import lombok.RequiredArgsConstructor;
+
 @Service
+@RequiredArgsConstructor
 public class AirplaneService {
     private final AirplaneRepository airplaneRepository;
+    private final FlightDao flightDao;
 
-    public AirplaneService(AirplaneRepository airplaneRepository) {
-        this.airplaneRepository = airplaneRepository;
-    }
-
-    public List<Airplane> findAllAirplanes() {
-        return airplaneRepository.findAll();
-    }
-
-    public Page<Airplane> findAllAirplanes(Integer pageIndex,
-            Integer pageSize) {
+    public Page<Airplane> findAll(Integer pageIndex, Integer pageSize) {
         Pageable pageable = PageRequest.of(pageIndex, pageSize);
         return airplaneRepository.findAll(pageable);
     }
 
-    public Page<Airplane> searchAirplanesPage(String term, Integer pageIndex,
+    public Page<Airplane> search(String term, Integer pageIndex,
             Integer pageSize) {
         Pageable pageable = PageRequest.of(pageIndex, pageSize);
         return airplaneRepository.findByModelContaining(term, pageable);
@@ -43,20 +37,22 @@ public class AirplaneService {
         return airplaneRepository.findDistinctByModelContaining(term, pageable);
     }
 
-    public Airplane findAirplaneById(Long id) {
+    public Airplane findById(Long id) {
         return airplaneRepository.findById(id)
                 .orElseThrow(() -> new AirplaneNotFoundException(id));
     }
 
-    public List<Airplane> findByModelContaining(String model) {
-        return airplaneRepository.findByModelContaining(model);
+    public Page<Airplane> findByModelContaining(String model, Integer pageIndex,
+            Integer pageSize) {
+        Pageable pageable = PageRequest.of(pageIndex, pageSize);
+        return airplaneRepository.findByModelContaining(model, pageable);
     }
 
-    public Airplane createAirplane(Airplane airplane) {
+    public Airplane create(Airplane airplane) {
         return airplaneRepository.save(airplane);
     }
 
-    public Airplane updateAirplane(Airplane airplane) {
+    public Airplane update(Airplane airplane) {
         airplaneRepository.findById(airplane.getId()).orElseThrow(
                 () -> new ResponseStatusException(HttpStatus.NOT_FOUND,
                         "Could not find airplane with id = "
@@ -64,10 +60,17 @@ public class AirplaneService {
         return airplaneRepository.save(airplane);
     }
 
-    public void deleteAirplaneById(Long id) throws ResponseStatusException {
+    public void deleteById(Long id) throws ResponseStatusException {
         airplaneRepository.findById(id).orElseThrow(
                 () -> new ResponseStatusException(HttpStatus.NOT_FOUND,
                         "Could not find airplane with id = " + id));
-        airplaneRepository.deleteById(id);
+        if (flightDao.areAnyActiveFlightsWithAirplane(id)) {
+            String reason = String.format(
+                    "Airplane with id: %s which has at least one active flight cannot be deleted.",
+                    id);
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, reason);
+        } else {
+            airplaneRepository.deleteById(id);
+        }
     }
 }
